--- mobile-app/constants/Colors.ts
@@ -1,41 +1,0 @@
-/**
- * Below are the colors that are used in the app. The colors are defined in the light and dark mode.
- * There are many other ways to style your app. For example, [Nativewind](https://www.nativewind.dev/), [Tamagui](https://tamagui.dev/), [unistyles](https://reactnativeunistyles.vercel.app), etc.
- */
-
-const tintColorLight = '#0a7ea4';
-const tintColorDark = '#fff';
-
-export const Colors = {
-  light: {
-    text: '#090A05',
-    lightText: '#C8C3BB',
-    mutedText: '#6C6F6F',
-    icon: '#090A05',
-    offWhite: '#F3F3F1',
-    brightBlue: '#0255EE',
-    mediumBlue: '#002484',
-    darkBlue: '#001756',
-    brightGreen: '#4B701F',
-    mediumGreen: '#395518',
-    darkGreen: '#2B4111',
-    brightYellow: '#FEC001',
-    mediumYellow: '#BF9C09',
-    darkYellow: '#886D00',
-    brightRed: '#FE3302',
-    mediumRed: '#CA1401',
-    darkRed: '#6A0000',
-    background: '#fff',
-    tint: tintColorLight,
-    tabIconDefault: '#687076',
-    tabIconSelected: tintColorLight,
-  },
-  dark: {
-    text: '#ECEDEE',
-    background: '#151718',
-    tint: tintColorDark,
-    icon: '#9BA1A6',
-    tabIconDefault: '#9BA1A6',
-    tabIconSelected: tintColorDark,
-  },
-};
